project/requirements.txt
PyQt6==6.7.1
faiss-cpu==1.8.0.post1
numpy>=1.24.0
sentence-transformers==3.0.1
rank-bm25==0.2.2
pypdf==4.3.1
pymupdf==1.24.9
pdfminer.six==20240706
chardet==5.2.0
docx2txt==0.8
python-pptx==1.0.2
pandas>=2.0.0
openpyxl>=3.1.0
openai>=1.40.0
transformers>=4.43.0
torch>=2.2.0
